- Removing unnecessary TyrusServletServerContainer which is not really used except to collect dynamically added endpoint configs and classes. - TyrusServerContainer is made to implement to ServerContainer and exposed using servlet context attribute
