+++ containers/servlet/src/main/java/org/glassfish/tyrus/servlet/TyrusServletFilter.java
@@ -42,22 +42,17 @@ package org.glassfish.tyrus.servlet;
 import java.io.IOException;
 import java.net.URI;
 import java.util.Enumeration;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
 import javax.websocket.server.ServerContainer;
-import javax.websocket.server.ServerEndpointConfig;
 
 import javax.servlet.Filter;
 import javax.servlet.FilterChain;
 import javax.servlet.FilterConfig;
-import javax.servlet.FilterRegistration;
-import javax.servlet.ServletContext;
 import javax.servlet.ServletException;
 import javax.servlet.ServletRequest;
 import javax.servlet.ServletResponse;
@@ -70,7 +65,6 @@ import javax.servlet.http.WebConnection;
 
 import org.glassfish.tyrus.core.RequestContext;
 import org.glassfish.tyrus.core.Utils;
-import org.glassfish.tyrus.server.ServerContainerFactory;
 import org.glassfish.tyrus.websockets.HandshakeException;
 import org.glassfish.tyrus.websockets.WebSocketEngine;
 
@@ -85,18 +79,9 @@ import org.glassfish.tyrus.websockets.WebSocketEngine;
  */
 public class TyrusServletFilter implements Filter, HttpSessionListener {
 
-    private static final int INFORMATIONAL_FIXED_PORT = 8080;
     private final static Logger LOGGER = Logger.getLogger(TyrusServletFilter.class.getName());
     private final WebSocketEngine engine = WebSocketEngine.getEngine();
     private org.glassfish.tyrus.server.TyrusServerContainer serverContainer = null;
-
-    private boolean registered = false;
-
-    // @ServerEndpoint annotated classes and classes extending ServerApplicationConfig
-    private Set<Class<?>> classes = null;
-    private ServletContext servletContext = null;
-    private final Set<Class<?>> dynamicallyDeployedClasses = new HashSet<Class<?>>();
-    private final Set<ServerEndpointConfig> dynamicallyDeployedServerEndpointConfigs = new HashSet<ServerEndpointConfig>();
 
     // I don't like this map, but it seems like it is necessary. I am forced to handle subscriptions
     // for HttpSessionListener because the listener itself must be registered *before* ServletContext
@@ -106,53 +91,15 @@ public class TyrusServletFilter implements Filter, HttpSessionListener {
     private final Map<HttpSession, TyrusHttpUpgradeHandler> sessionToHandler =
             new ConcurrentHashMap<HttpSession, TyrusHttpUpgradeHandler>();
 
-    public TyrusServletFilter() {
-    }
-
-    void addClass(Class<?> clazz) {
-        if (this.serverContainer != null) {
-            throw new IllegalStateException("Filter already initiated.");
-        }
-        this.dynamicallyDeployedClasses.add(clazz);
-
-        checkFilterRegistration();
-    }
-
-    void addServerEndpointConfig(ServerEndpointConfig serverEndpointConfig) {
-        if (this.serverContainer != null) {
-            throw new IllegalStateException("Filter already initiated.");
-        }
-        this.dynamicallyDeployedServerEndpointConfigs.add(serverEndpointConfig);
-
-        checkFilterRegistration();
-    }
-
-    private void checkFilterRegistration() {
-        if (servletContext != null && !registered) {
-            registered = true;
-
-            final FilterRegistration.Dynamic reg = servletContext.addFilter("WebSocket filter", this);
-            reg.setAsyncSupported(true);
-            reg.addMappingForUrlPatterns(null, true, "/*");
-            LOGGER.info("Registering WebSocket filter for url pattern /*");
-        }
-    }
-
     @Override
     public void init(FilterConfig filterConfig) throws ServletException {
-        String contextRoot = filterConfig.getServletContext().getContextPath();
-        this.serverContainer = ServerContainerFactory.create(ServletContainer.class, contextRoot, INFORMATIONAL_FIXED_PORT, classes, dynamicallyDeployedClasses, dynamicallyDeployedServerEndpointConfigs);
+        this.serverContainer = (org.glassfish.tyrus.server.TyrusServerContainer)filterConfig.getServletContext().getAttribute(ServerContainer.class.getName());
         try {
             serverContainer.start();
         } catch (Exception e) {
             throw new ServletException("Web socket server initialization failed.", e);
         } finally {
-
-            // remove reference to filter.
-            final ServerContainer container = (ServerContainer) filterConfig.getServletContext().getAttribute(TyrusServletServerContainer.SERVER_CONTAINER_ATTRIBUTE);
-            if (container instanceof TyrusServletServerContainer) {
-                ((TyrusServletServerContainer) container).doneDeployment();
-            }
+            serverContainer.doneDeployment();
         }
     }
 
@@ -314,21 +261,4 @@ public class TyrusServletFilter implements Filter, HttpSessionListener {
         serverContainer.stop();
     }
 
-    /**
-     * Set the {@link ServletContext}.
-     *
-     * @param servletContext to be set.
-     */
-    public void setServletContext(ServletContext servletContext) {
-        this.servletContext = servletContext;
-    }
-
-    /**
-     * Set the scanned classes.
-     *
-     * @param classes scanned classes.
-     */
-    public void setClasses(Set<Class<?>> classes) {
-        this.classes = classes;
-    }
 }
