+++ server/src/main/java/org/glassfish/tyrus/server/TyrusServerContainer.java
@@ -53,8 +53,8 @@ import javax.websocket.Endpoint;
 import javax.websocket.EndpointConfig;
 import javax.websocket.Extension;
 import javax.websocket.Session;
-import javax.websocket.WebSocketContainer;
 import javax.websocket.server.ServerApplicationConfig;
+import javax.websocket.server.ServerContainer;
 import javax.websocket.server.ServerEndpointConfig;
 
 import org.glassfish.tyrus.core.AnnotatedEndpoint;
@@ -72,18 +72,21 @@ import org.glassfish.tyrus.spi.TyrusServer;
  * @author Pavel Bucek (pavel.bucek at oracle.com)
  * @author Stepan Kopriva (stepan.kopriva at oracle.com)
  */
-public class TyrusServerContainer extends BaseContainer implements WebSocketContainer {
+public class TyrusServerContainer extends BaseContainer implements ServerContainer {
     private final TyrusServer server;
     private final String contextPath;
-    private final ServerApplicationConfig configuration;
+    private final Set<Class<?>> dynamicallyAddedClasses;
+    private final Set<ServerEndpointConfig> dynamicallyAddedEndpointConfigs;
     private final Set<SPIRegisteredEndpoint> endpoints = new HashSet<SPIRegisteredEndpoint>();
     private final ErrorCollector collector;
     private final ComponentProviderService componentProvider;
+    private final Set<Class<?>> classes;
 
     private long defaultMaxSessionIdleTimeout = 0;
     private long defaultAsyncSendTimeout = 0;
     private int maxTextMessageBufferSize = Integer.MAX_VALUE;
     private int maxBinaryMessageBufferSize = Integer.MAX_VALUE;
+    private boolean canDeploy = true;
 
     /**
      * Create new {@link TyrusServerContainer}.
@@ -103,8 +106,9 @@ public class TyrusServerContainer extends BaseContainer implements WebSocketCont
         this.collector = new ErrorCollector();
         this.server = server;
         this.contextPath = contextPath;
-        this.configuration = new TyrusServerConfiguration((classes == null ? Collections.<Class<?>>emptySet() : classes),
-                dynamicallyAddedClasses, dynamicallyAddedEndpointConfigs, this.collector);
+        this.classes = classes;
+        this.dynamicallyAddedClasses = dynamicallyAddedClasses;
+        this.dynamicallyAddedEndpointConfigs = dynamicallyAddedEndpointConfigs;
         this.componentProvider = ComponentProviderService.create();
     }
 
@@ -115,6 +119,9 @@ public class TyrusServerContainer extends BaseContainer implements WebSocketCont
      * @throws DeploymentException when any deployment related error is found; should contain list of all found issues.
      */
     public void start() throws IOException, DeploymentException {
+        ServerApplicationConfig configuration = new TyrusServerConfiguration((classes == null ? Collections.<Class<?>>emptySet() : classes),
+                dynamicallyAddedClasses, dynamicallyAddedEndpointConfigs, this.collector);
+
         // start the underlying server
         server.start();
         try {
@@ -229,4 +236,27 @@ public class TyrusServerContainer extends BaseContainer implements WebSocketCont
     public void setDefaultMaxSessionIdleTimeout(long defaultMaxSessionIdleTimeout) {
         this.defaultMaxSessionIdleTimeout = defaultMaxSessionIdleTimeout;
     }
+
+    @Override
+    public void addEndpoint(Class<?> endpointClass) throws DeploymentException {
+        if (canDeploy) {
+            dynamicallyAddedClasses.add(endpointClass);
+        } else {
+            throw new IllegalStateException("Not in 'deploy' scope.");
+        }
+    }
+
+    @Override
+    public void addEndpoint(ServerEndpointConfig serverConfig) throws DeploymentException {
+        if (canDeploy) {
+            dynamicallyAddedEndpointConfigs.add(serverConfig);
+        } else {
+            throw new IllegalStateException("Not in 'deploy' scope.");
+        }
+    }
+
+    public void doneDeployment() {
+        canDeploy = false;
+    }
+
 }
